AiAdaptiveQuizBackend/AiAdaptiveQuizBackend/urls.py
from django.contrib import admin
from django.urls import path, include
from django.conf import settings
from django.conf.urls.static import static
from drf_yasg.views import get_schema_view
from drf_yasg import openapi
from rest_framework import permissions

schema_view = get_schema_view(
    openapi.Info(
        title=settings.SITE_NAME,
        default_version=settings.SITE_VERSION,
        description=f"{settings.SITE_NAME} API's",
        terms_of_service="https://www.google.com/policies/terms/",
        contact=openapi.Contact(email="venkatnvs2005@gmail.com"),
        license=openapi.License(name="BSD License"),
    ),
    public=True,
    permission_classes=(permissions.AllowAny,),
)

urlpatterns = [
    path('admin/', admin.site.urls),
    path('api/auth/', include('account.urls')),
    path('api/core/', include('core.urls')),
    path('', schema_view.with_ui('swagger', cache_timeout=0), name='schema-swagger-ui'),
    path('redoc/', schema_view.with_ui('redoc', cache_timeout=0), name='schema-redoc'),
]

if settings.DEBUG:
    urlpatterns += static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
    urlpatterns += static(settings.MEDIA_URL,document_root=settings.MEDIA_ROOT)
    
admin.site.site_title = "AdaptiveLearn AI site admin"
admin.site.site_header = "AdaptiveLearn AI administration"
admin.site.index_title = "Site administration"