AiAdaptiveQuizBackend/requirements.txt
Django
django-decouple
djangorestframework
djangorestframework-simplejwt
django-cors-headers
validate_email
drf-yasg
pillow
openai-whisper
pytube
yt-dlp